avoid stat64 redifinition with musl
musl define stat64() as an alias to stat(), this leads to a
stat() function redefinition.

On Glibc stat64() is an alias to __xstat64() which is has an
hidden prototype, that is why stat64() can be defined in padsp.c.

stat64 can't be undef since it's also the name of the type
"struct stat64".

Rename padsp stat64() to _stat64(), also rename the function
pointer _stat64 to __stat64.

Fixes:
In file included from utils/padsp.c:40:0:
utils/padsp.c:2516:5: error: redefinition of ‘stat’
 int stat64(const char *pathname, struct stat64 *buf) {

Signed-off-by: Romain Naour <romain.naour@openwide.fr>